+++ sphinx/source/changelog.rst
@@ -24,6 +24,7 @@ Changelog
   memory and I/O contention and the MEMPTR register)
 * Added support for multiple colours in the border area of the screen displayed
   by :ref:`rzxplay.py` and :ref:`trace.py`
+* Made the :ref:`AY audio writer <ayAudioWriter>` a pluggable component
 * Fixed how :ref:`rzxinfo.py` detects a signed RZX file and encrypted frames
 * Fixed the bug in :ref:`sna2ctl.py` that enables the :ref:`rstHandler` when
   the :ref:`comment generator <commentGenerator>` is enabled